ai-service/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
langchain
langchain-openai
pytesseract
pdfplumber
Pillow

